requirements.txt
fastapi
jinja2
uvicorn
aio_pika
python-dotenv